[app] migrate to compose destinations
This is a fairly large changes with some minor regressions
but generally massively simplifies navigation infrastructure around the application and sets up easier bottom
nav in the future too.

The main impact is that is the architecture now slightly favours passing state (improvToolsAppState instead of
callbacks. future work will fix this. ITAS is really a broken model in general...